refactor: EntityMenu styles to a separate file
