+++ wondrous-app/components/Common/Sidebar/Entity/AboutEntity/EntityMenu/styles.tsx
@@ -1,20 +1,17 @@
 import { ButtonBase, Menu, MenuItem, Typography } from '@mui/material';
-import { SafeImage } from 'components/Common/Image';
 import PodsIcon from 'components/Common/Sidebar/Common/icons/pods.svg';
 import Arrow from 'components/Icons/arrow.svg';
 import { DAOIcon } from 'components/Icons/dao';
-import { useState } from 'react';
 import styled from 'styled-components';
-import { useBoards } from 'utils/hooks';
 
-const ButtonIcon = styled.div`
+export const ButtonIcon = styled.div`
   width: 28px;
   height: 28px;
   background: ${({ theme }) => theme.palette.grey78};
   border-radius: 6px;
 `;
 
-const Text = styled(Typography)`
+export const Text = styled(Typography)`
   && {
     font-family: ${({ theme }) => theme.typography.fontFamily};
     font-weight: 500;
@@ -26,7 +23,7 @@ const Text = styled(Typography)`
   }
 `;
 
-const Button = styled(ButtonBase)`
+export const Button = styled(ButtonBase)`
   && {
     max-width: fit-content;
     padding: 2px;
@@ -63,7 +60,7 @@ export const ArrowIcon = styled((props) => (
   }
 `;
 
-const MenuStyled = styled(Menu)`
+export const MenuStyled = styled(Menu)`
   && {
     margin-top: 6px;
     .MuiMenu-paper {
@@ -81,7 +78,7 @@ const MenuStyled = styled(Menu)`
   }
 `;
 
-const Item = styled(MenuItem)`
+export const Item = styled(MenuItem)`
   && {
     font-family: ${({ theme }) => theme.typography.fontFamily};
     font-weight: 500;
@@ -130,44 +127,3 @@ export const NoLogoPod = styled((props) => (
   justify-content: center;
   background: ${({ theme }) => theme.palette.grey87}; ;
 `;
-
-const EntityMenu = ({ name, id, router, thumbnailPicture, profilePicture, canManage }) => {
-  const { orgBoard } = useBoards();
-  const [anchorEl, setAnchorEl] = useState(null);
-  const open = Boolean(anchorEl);
-  const handleClick = (event) => setAnchorEl(event.currentTarget);
-  const handleClose = () => setAnchorEl(null);
-  const handleOnClickNotifications = () =>
-    router.push(orgBoard ? `/organization/settings/${id}/notifications` : `/pod/settings/${id}/notifications`);
-  const NoLogo = orgBoard ? NoLogoDAO : NoLogoPod;
-  return (
-    <>
-      <Button onClick={handleClick} open={open} disabled={!canManage}>
-        <ButtonIcon>
-          {thumbnailPicture || profilePicture ? (
-            <SafeImage
-              useNextImage={false}
-              src={thumbnailPicture || profilePicture}
-              width={28}
-              height={28}
-              objectFit="cover"
-              style={{
-                borderRadius: '2px',
-              }}
-            />
-          ) : (
-            <NoLogo />
-          )}
-        </ButtonIcon>
-        <Text>{name}</Text>
-        {canManage && <ArrowIcon open={open} />}
-      </Button>
-      <MenuStyled anchorEl={anchorEl} open={open} onClose={handleClose}>
-        <Item onClick={handleOnClickNotifications}>Notification Settings</Item>
-        {/* <Item>Leave Project</Item> NOTE: There's no endpoint for this yet. */}
-      </MenuStyled>
-    </>
-  );
-};
-
-export default EntityMenu;
